SWC-1364: fix for large file upload

+++ src/main/java/org/sagebionetworks/web/client/SynapseJSNIUtils.java
@@ -47,7 +47,7 @@ public interface SynapseJSNIUtils {
 
 	public boolean isDirectUploadSupported();
 	
-	public void uploadFileChunk(String contentType, String fileFieldId, int startByte, int endByte, String url, XMLHttpRequest xhr, ProgressCallback callback);
+	public void uploadFileChunk(String contentType, String fileFieldId, double startByte, double endByte, String url, XMLHttpRequest xhr, ProgressCallback callback);
 	
 	public String getContentType(String fileFieldId);
